Fix a problem with DC offset when muting a channel (issue #4). Disable frequency limiter if the internal rate converter is not active (i.e. quality == 0).

+++ emu2149.h
@@ -24,7 +24,7 @@ extern "C"
     uint8_t quality;
     uint8_t clk_div;
 
-    uint32_t count[3];
+    uint16_t count[3];
     uint8_t volume[3];
     uint16_t freq[3];
     uint8_t edge[3];
